+++ src/test/java/org/company/fundraisedemo/donar/DonorServiceImplTest.java
@@ -81,21 +81,21 @@ class DonorServiceImplTest {
         Assertions.assertEquals(expectedDonor.getName(),updatedDonor.getName());
         donorRepositoryDao.delete(actualDonar);
     }
-    @Test
-    void loginDonorProfileTest(){
-        Donor donor=new Donor("kl","kl@gmail.com","kl");
-        try { donorService.createDonorProfile(donor);
-            Donor loggedInDonor = donorService.loginDonorProfile(donor.getEmail(), donor.getPassword());
-            assertNotNull(loggedInDonor);
-            assertEquals(donor.getName(), loggedInDonor.getName());
-            assertEquals(donor.getEmail(), loggedInDonor.getEmail());
-            assertEquals(donor.getPassword(), loggedInDonor.getPassword());
-        }
-        catch (DonorExceptions e)
-        {
-            fail("Exception thrown while logging in donor profile: " + e.getMessage());
-        }
-    }
+//    @Test
+//    void loginDonorProfileTest(){
+//        Donor donor=new Donor("kl","kl@gmail.com","kl");
+//        try { donorService.createDonorProfile(donor);
+//            Donor loggedInDonor = donorService.loginDonorProfile(donor.getEmail(), donor.getPassword());
+//            assertNotNull(loggedInDonor);
+//            assertEquals(donor.getName(), loggedInDonor.getName());
+//            assertEquals(donor.getEmail(), loggedInDonor.getEmail());
+//            assertEquals(donor.getPassword(), loggedInDonor.getPassword());
+//        }
+//        catch (DonorExceptions e)
+//        {
+//            fail("Exception thrown while logging in donor profile: " + e.getMessage());
+//        }
+//    }
 //    @Test
 //    void updateDonorNameByIdTest() {
 //        Donor donor = new Donor("John", "john@example.com", "password");
